requirements.txt
fastapi
uvicorn
tensorflow
numpy
opencv-python
Pillow
matplotlib
python-multipart